v10.5.2

+++ Win/Samples/LoopThroughWithOpenGLCompositing/OpenGLComposite.cpp
@@ -780,7 +780,7 @@ ULONG STDMETHODCALLTYPE		PinnedMemoryAllocator::Release(void)
 }
 
 // IDeckLinkMemoryAllocator methods
-HRESULT STDMETHODCALLTYPE	PinnedMemoryAllocator::AllocateBuffer (unsigned long bufferSize, void* *allocatedBuffer)
+HRESULT STDMETHODCALLTYPE	PinnedMemoryAllocator::AllocateBuffer (unsigned int bufferSize, void* *allocatedBuffer)
 {
 	if (mFrameCache.empty())
 	{
